split DenoInspector
