+++ runtime/inspector_session.rs
@@ -12,100 +12,10 @@ use deno_core::serde_json::json;
 use deno_core::serde_json::Value;
 use deno_core::v8;
 use std::collections::HashMap;
-use std::mem::MaybeUninit;
 use std::ops::Deref;
 use std::ops::DerefMut;
-use std::ptr;
-use std::cell::RefCell;
-
-pub struct DenoInspector {
-  v8_inspector_client: v8::inspector::V8InspectorClientBase,
-  v8_inspector: v8::UniquePtr<v8::inspector::V8Inspector>,
-  flags: RefCell<InspectorFlags>,
-}
-
-impl Deref for DenoInspector {
-  type Target = v8::inspector::V8Inspector;
-  fn deref(&self) -> &Self::Target {
-    self.v8_inspector.as_ref().unwrap()
-  }
-}
-
-impl DerefMut for DenoInspector {
-  fn deref_mut(&mut self) -> &mut Self::Target {
-    self.v8_inspector.as_mut().unwrap()
-  }
-}
-
-impl v8::inspector::V8InspectorClientImpl for DenoInspector {
-  fn base(&self) -> &v8::inspector::V8InspectorClientBase {
-    &self.v8_inspector_client
-  }
-
-  fn base_mut(&mut self) -> &mut v8::inspector::V8InspectorClientBase {
-    &mut self.v8_inspector_client
-  }
-
-  fn run_message_loop_on_pause(&mut self, context_group_id: i32) {
-    assert_eq!(context_group_id, DenoInspector::CONTEXT_GROUP_ID);
-    self.flags.borrow_mut().on_pause = true;
-  }
-
-  fn quit_message_loop_on_pause(&mut self) {
-    self.flags.borrow_mut().on_pause = false;
-  }
-
-  fn run_if_waiting_for_debugger(&mut self, context_group_id: i32) {
-    assert_eq!(context_group_id, DenoInspector::CONTEXT_GROUP_ID);
-    self.flags.borrow_mut().session_handshake_done = true;
-  }
-}
-
-impl DenoInspector {
-  const CONTEXT_GROUP_ID: i32 = 1;
-
-  pub fn new(js_runtime: &mut deno_core::JsRuntime) -> Box<Self> {
-    let context = js_runtime.global_context();
-    let scope = &mut v8::HandleScope::new(js_runtime.v8_isolate());
-
-    // Create DenoInspector instance.
-    let mut self_ = new_box_with(|self_ptr| {
-      let v8_inspector_client =
-        v8::inspector::V8InspectorClientBase::new::<Self>();
-
-      let flags = InspectorFlags::new();
-
-      Self {
-        v8_inspector_client,
-        v8_inspector: Default::default(),
-        flags,
-      }
-    });
-    self_.v8_inspector =
-      v8::inspector::V8Inspector::create(scope, &mut *self_).into();
-
-    // Tell the inspector about the global context.
-    let context = v8::Local::new(scope, context);
-    let context_name = v8::inspector::StringView::from(&b"global context"[..]);
-    self_.context_created(context, Self::CONTEXT_GROUP_ID, context_name);
-
-    self_
-  }
-}
-
-#[derive(Default)]
-struct InspectorFlags {
-  waiting_for_session: bool,
-  session_handshake_done: bool,
-  on_pause: bool,
-}
-
-impl InspectorFlags {
-  fn new() -> RefCell<Self> {
-    let self_ = Self::default();
-    RefCell::new(self_)
-  }
-}
+use crate::deno_inspector::DenoInspector;
+use crate::deno_inspector::new_box_with;
 
 /// A local inspector session that can be used to send and receive protocol messages directly on
 /// the same thread as an isolate.
@@ -249,11 +159,4 @@ impl InspectorSession {
     let result = response.get("result").unwrap().clone();
     Ok(result)
   }
-}
-
-fn new_box_with<T>(new_fn: impl FnOnce(*mut T) -> T) -> Box<T> {
-  let b = Box::new(MaybeUninit::<T>::uninit());
-  let p = Box::into_raw(b) as *mut T;
-  unsafe { ptr::write(p, new_fn(p)) };
-  unsafe { Box::from_raw(p) }
 }
